Split up gigantic module
